requirements.txt
bleach==6.0.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.6
cryptography==41.0.3
Deprecated==1.2.14
docutils==0.20.1
Flask==2.3.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
keyring==24.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
more-itertools==9.1.0
Opensource-License-Compatibility-from-DoWell==0.3.0
pkginfo==1.9.6
pycparser==2.21
PyGithub==1.59.1
Pygments==2.15.1
PyJWT==2.8.0
PyNaCl==1.5.0
readme-renderer==40.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.4.2
six==1.16.0
twine==4.0.2
urllib3==2.0.3
webencodings==0.5.1
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.16.0
